% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{get_bb}
\alias{get_bb}
\title{Retrieve a single bounding box as either a table, vector of sf POLYGON}
\usage{
get_bb(reg, form = c("table", "bb", "sf")[2])
}
\arguments{
\item{reg}{character, one or more named regions to retrieve.
By default 'world'.  Use all to retrieve all known bbs.}

\item{form}{character, either 'bb', 'table' or 'sf'}
}
\value{
4-element numeric vector, or a list of 4-element vectors,
        a table (tibble) or sf POLYGON table
}
\description{
Retrieve a single bounding box as either a table, vector of sf POLYGON
}
