% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{get_table}
\alias{get_table}
\title{Retrieve the entire table of known bounding boxes}
\usage{
get_table(filename = system.file("bbox_lonlat.csv", package = "cofbb"))
}
\value{
table of bounding boxes
}
\description{
Retrieve the entire table of known bounding boxes
}
