% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bb}
\alias{plot_bb}
\title{Plot one or more bounding boxes}
\usage{
plot_bb(x)
}
\arguments{
\item{x}{sf data frame of bounding boxes}
}
\description{
Suggests leaflet package, but will do simple plot if not available
}
