% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_as_POLYGON}
\alias{bb_as_POLYGON}
\title{Convert a 4 element bb vector into a polygon}
\usage{
bb_as_POLYGON(x)
}
\arguments{
\item{x}{4 element numeric bounding box}
}
\value{
\code{POLYGON} object
}
\description{
Convert a 4 element bb vector into a polygon
}
