% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{get_bb}
\alias{get_bb}
\title{Retrieve a single bounding box as either a table or vector}
\usage{
get_bb(reg, form = c("table", "bb")[2])
}
\arguments{
\item{reg}{character, one or more named regions to retrieve.
By default 'world'}

\item{form}{character, either 'bb' or 'table'}
}
\value{
4-element numeric vector, or a list of 4-element vectors
        or a table (tibble)
}
\description{
Retrieve a single bounding box as either a table or vector
}
